+++ core/src/main/kotlin/org/cryptobiotic/rlauxe/core/AlphaMart.kt
@@ -130,65 +130,3 @@ class AlphaMart(
         return TestH0Result(status, sampleNumber, sampleMean, pvalues, etajs, mjs)
     }
 }
-
-// 3. Pseudo-algorithm for ballot-level comparison and ballot-polling audits
-// • Set audit parameters:
-//  – Select the risk limit α ∈ (0, 1)
-//  - decide whether to sample with or without replacement.
-//  – Set upper as appropriate for the assertion under audit.
-//  – Set N to the number of ballot cards in the population of cards from which the sample is drawn.
-// (the rest is all in the estimFn)
-//  – Set η0
-//    For polling audits, η0 could be the reported mean value of the assorter.
-//	    For instance, for the assertion corresponding to checking whether w got more votes than ℓ,
-//	      η0 = (Nw + Nc /2)/N , where Nw is the number of votes reported for w , Nℓ is the
-//	   number of votes reported for ℓ, and Nc = N − Nw − Nℓ is the number of ballot cards
-//	   reported to have a vote for some other candidate or no valid vote in the contest.
-//     For comparison audits, η0 can be based on assumed or historical rates of overstatement errors.
-//
-//  – Define the function to update η based on the sample,
-//	  e.g, η(i, X i−1 ) = ((d * η0 + S)/(d + i − 1) ∨ (eps(i) + µi )) ∧ u,    (2.5.2, eq 14, "truncated shrinkage")
-//	     where S = Sum i−1 k=1 (Xk) is the sample sum of the first i − 1 draws
-//	     and eps(i) = c/ sqrt(d + i − 1)
-//	  set any free parameters in the function (e.g., d and c in this example). The only requirement is that
-//	     η(i, X i−1 ) ∈ (µi , u), where µi := E(Xi |X i−1 ) is computed under the null.
-
-// simple version of alpha_mart from ONEAUDIT
-// def alpha_mart(x: np.array, N: int, mu: float=1/2, eta: float=1-np.finfo(float).eps, u: float=1,
-//               estim: callable=shrink_trunc) -> np.array :
-//    '''
-//    Finds the ALPHA martingale for the hypothesis that the population
-//    mean is less than or equal to t using a martingale method,
-//    for a population of size N, based on a series of draws x.
-//
-//    The draws must be in random order, or the sequence is not a martingale under the null
-//
-//    If N is finite, assumes the sample is drawn without replacement
-//    If N is infinite, assumes the sample is with replacement
-//
-//    Parameters
-//    ----------
-//    x : list corresponding to the data
-//    N : int
-//        population size for sampling without replacement, or np.infinity for sampling with replacement
-//    mu : float in (0,1)
-//        hypothesized fraction of ones in the population
-//    eta : float in (t,1)
-//        alternative hypothesized population mean
-//    estim : callable
-//        estim(x, N, mu, eta, u) -> np.array of length len(x), the sequence of values of eta_j for ALPHA
-//
-//    Returns
-//    -------
-//    terms : array
-//        sequence of terms that would be a nonnegative supermartingale under the null
-//    '''
-//    S = np.insert(np.cumsum(x),0,0)[0:-1]  # 0, x_1, x_1+x_2, ...,
-//    j = np.arange(1,len(x)+1)              # 1, 2, 3, ..., len(x)
-//    m = (N*mu-S)/(N-j+1) if np.isfinite(N) else mu   # mean of population after (j-1)st draw, if null is true
-//    etaj = estim(x, N, mu, eta, u)
-//    with np.errstate(divide='ignore',invalid='ignore'):
-//        term = (x*etaj/m + (u-x)*(u-etaj)/(u-m))/u
-//        terms = np.cumprod((x*etaj/m + (u-x)*(u-etaj)/(u-m))/u)
-//    terms[m<0] = np.inf
-//    return terms
