+++ core/src/test/data/corla/Boulder2023_Test_Canonical_List.csv
@@ -1,0 +1,39 @@
+CountyName,ContestName,ContestChoices
+Boulder,City of Boulder Mayoral,"Aaron Brockett,Nicole Speer,Bob Yates,Paul Tweedlie,"
+Boulder,City of Boulder Council,"Terri Brncic,Jenny Robins,Aaron Gabriel Neyer,Jacques Decalo,Silas Atkins,Waylon Lewis,Ryan Schuchard,Tara Winer,Tina Marquis,Taishya Adams"
+Boulder,City of Lafayette City Council,"Tim Barnes,JD Mangat,Eric Ryant,John W. Watson,Gala W. Orba,David Fridland,Crystal Gallegos"
+Boulder,City of Longmont - Mayor,"Ethan Augreen,Joan Peck,Terri Goon"
+Boulder,City of Longmont - City Council Member At-Large,"Sean P. McCoy,Steve Altschuler,Beka Venturella"
+Boulder,City of Longmont - Council Member Ward 1,"Nia Wassink,Diane Crist,Harrison Earl"
+Boulder,City of Longmont - Council Member Ward 3,"Ron Gallegos,Gary Hodges,Susie Hidalgo-Fahring,Spencer Adams"
+Boulder,City of Louisville Mayor At-Large (4 Year Term),"Sherry Sommer,Chris Leh,Josh Cooperman"
+Boulder,City of Louisville City Council Ward 1 (4-year term),"J. Caleb Dickinson"
+Boulder,City of Louisville City Council Ward 2 (4-year term),"Deborah Fahey,George Colbert"
+Boulder,City of Louisville City Council Ward 3,"Dietrich Hoefner,Barbara Hamlington"
+Boulder,Boulder Valley School District RE-2 Director District A (4 Years),"Jason Unger,Neil Fishman"
+Boulder,Boulder Valley School District RE-2 Director District C (4 Years),"Andrew Steffl,Alex Medler,Cynthia Nevison"
+Boulder,Boulder Valley School District RE-2 Director District D (4 Years),"Andrew Brandt,Lalenia Quinlan Aweida"
+Boulder,Boulder Valley School District RE-2 Director District G (4 Years),"Anil Kiran Pesaramelli,Stuart Lord,Jorge Chavez"
+Boulder,Estes Park School District R-3 School Board Director At Large (4 Year),"Kevin G. Morris,Kyri Cox,Brenda L. Wyss,Brad Shochat"
+Boulder,Thompson R2-J School District Board of Education Director District A (4 Year Term),"Ryan Wilcken,Dawn Kirk"
+Boulder,Thompson R2-J School District Board of Education Director District C (4 Year Term),"Nancy Rumfelt,Briah Freeman"
+Boulder,Thompson R2-J School District Board of Education Director District D (4 Year Term),"Denise Alvine Chapman,Yazmin Navarro"
+Boulder,Thompson R2-J School District Board of Education Director District G (4 Year Term),"Stu Boyd,Elizabeth Kearney"
+Boulder,City of Longmont Municipal Court Judge - Frick,"Yes/For,No/Against"
+Boulder,Proposition HH (Statutory),"Yes/For,No/Against"
+Boulder,Proposition II (Statutory),"Yes/For,No/Against"
+Boulder,Boulder County Ballot Issue 1A,"Yes/For,No/Against"
+Boulder,Boulder County Ballot Issue 1B,"Yes/For,No/Against"
+Boulder,City of Boulder Ballot Issue 2A,"Yes/For,No/Against"
+Boulder,City of Boulder Ballot Question 2B,"Yes/For,No/Against"
+Boulder,City of Boulder Ballot Question 302,"Yes/For,No/Against"
+Boulder,Town of Erie Ballot Question 3A,"Yes/For,No/Against"
+Boulder,Town of Erie Ballot Question 3B,"Yes/For,No/Against"
+Boulder,City of Longmont Ballot Issue 3C,"Yes/For,No/Against"
+Boulder,City of Longmont Ballot Issue 3D,"Yes/For,No/Against"
+Boulder,City of Longmont Ballot Issue 3E,"Yes/For,No/Against"
+Boulder,City of Louisville Ballot Issue 2C,"Yes/For,No/Against"
+Boulder,Town of Superior Ballot Question 301,"Yes/For,No/Against"
+Boulder,Town of Superior - Home Rule Charter Commission,"Dalton Valette,Heather Cracraft,Ryan Hitchler,Claire Dixon,Ryan Welch,Jeff Chu,Sean Maday,Clint Folsom,Chris Hanson,Stephanie Schader,Mike Foster"
+Boulder,Nederland Eco Pass Public Improvement District Ballot Issue 6A,"Yes/For,No/Against"
+Boulder,North Metro Fire Rescue District Ballot Issue 7A,"Yes/For,No/Against"
